+++ lib/src/tools/error_reporter_for_native.dart
@@ -20,8 +20,8 @@ class ErrorReporterForNative {
 
   ErrorReporterForNative._internal();
 
-  late String _endpointUrl;
-  late String _appVersion;
+  String? _endpointUrl;
+  String? _appVersion;
   Map<String, dynamic>? _extraInfo;
 
   Duration _rateLimitWindow = Duration(seconds: 60);
@@ -36,6 +36,8 @@ class ErrorReporterForNative {
 
   // Flag to allow reporting.
   bool allowReporting = true;
+
+  bool _isInitialized = false;
 
   /// (en)　Initialize this class. Run this after calling
   /// WidgetsFlutterBinding.ensureInitialized(); in main.dart.
@@ -171,6 +173,7 @@ class ErrorReporterForNative {
         }
       }
     };
+    _isInitialized = true;
   }
 
   /// (en) Sends the error content to the backend.
@@ -236,6 +239,11 @@ class ErrorReporterForNative {
         return;
       }
 
+      if (!_isInitialized) {
+        debugPrint('[ErrorReporter] Requires running init.');
+        return;
+      }
+
       // エラーレポートが無限ループなどにならないように、
       // 指定時間中の送信上限を超えたら無視する設定。
       final now = DateTime.now();
@@ -275,7 +283,7 @@ class ErrorReporterForNative {
 
       // 送信データを設定。
       final Map<String, dynamic> reportData = ErrorReportObj(
-              _appVersion,
+              _appVersion!,
               error.toString(),
               stackTrace?.toString(),
               now.toIso8601String(),
@@ -286,7 +294,7 @@ class ErrorReporterForNative {
       // バックエンドに送信。
       try {
         final response = await UtilHttpsForNative.post(
-            _endpointUrl, reportData, EnumPostEncodeType.json,
+            _endpointUrl!, reportData, EnumPostEncodeType.json,
             jwt: jwt,
             badCertificateCallback: badCertificateCallback,
             connectionTimeout: connectionTimeout,